fix: try to address an issue with client-side routing due to regex changes
I've been seeing issues with client-side routing, where it would not match paths. It looks like there have been some recent changes in this area, and if I make the leading `/` in this regex optional, things start working again.

I can't claim to totally understand what's going on here, but this seems to fix the weird issues I was seeing.