feat(adpater-node): support all esbuild build options
This PR allows to change the options passed to esbuild build which allows  to customize the server generation like target node version, sourcemap, inject and more.

It contains one braking change: `out` is renamed to `outdir` to follow the esbuild API. It is possible to avoid a breaking change by making `out` an alias for `outdir`. I'm happy to adjust the PR.