adapter-vercel should respect esbuild configs
This patch attempts to solve https://github.com/sveltejs/kit/issues/1706
We pass through the esbuild configurations in svelte.config.cjs into the
esbuild step of the vercel adapter.

This is primarily to solve the case where we have external packages that
are skipped by esbuild in adapater-node, but not adpater-vercel.