Update example config with correct defaults
The `svelte.config.js` example in the README included `host: 'HOST'` and `port: 'PORT'`
and a comment that says `default options are shown`. This implies that if `HOST` or `PORT`
environment variables must be specified. But then below it says
`By default, the server will accept connections on `0.0.0.0` using port 3000.`