Add support for Async Iterator endpoints
This commit demonstrates the ability to easily add Async Iterator
SvelteKit Endpoints. By having an endpoint return an async iterator as
the body, SvelteKit will keep emitting writes for as long as the
endpoint yields values. Or, for as long as the client holds the
request open; which ever ends first.