[fix] error components render with correct layout
If a deeply-nested page produced an error in its load() function, and
that page had multiple layout components but only a top-level error
component, the error component rendered differently in server-side and
client-side rendering. Server-side rendering correctly chose the
top-level layout component to render a top-level error component, but
client-side rendering was incorrectly choosing the layout of the
deeply-nested page that caused the error, instead of the layout of the
folder the error component belonged to.

With this commit, client-side rendering now uses the same approach as
server-side rendering to choose layouts for error components.