[fix] deeply-nested error components render right
Under some circumstances, a deeply-nested error component could render
with a layout from a parent folder instead of the layout from the error
component's own folder. This commit fixes that bug, so that error
components are correctly matched up with layout components in the same
folder even when deeply nested.