[fix] encode UInt8Array response bodies as base64
If an endpoint returns a UInt8Array body (e.g. for binary content), the
function handler will fail as a Netlify function response body needs to
be either a string or base64 encoded.

Let's encode UInt8Array response bodies as base64 so that endpoints
returning UInt8Arrays function properly with the Netlify adapter.