Bumps @cloudflare/kv-asset-handler
We've recently run into a bug serving static HTML pages alongside our
Svelte app (legacy codebase, working on removing these pages but it
takes time) with the following characteristics:

1. First load of a page to behave/run JS one way
2. Refresh/second load of the page using cached assets behave a
   completely different way

This bug appears to only affect webkit desktop browsers, and the page
works again if you empty cache and hard reload. We've been able to serve
the exact same assets from a simpler Cloudflare worker and everything
works as expected, and we've verified that:

1. LocalStorage values are not the culprit
2. Cookies are not used
3. Session state is not used
4. Headers are exactly the same, in Chrome devtools
5. Responses codes and values are exactly the same, in Chrome devtools

Manually updating this package appears to fix things, and we're not
exactly sure why yet, but it seemed like it would be a good thing to do
upstream so that everyone is on the latest. Still digging to try to find
a root cause.