Make ReadOnlyFormData getters work like FormData's
Previously, get and getAll would return the value with the specified key
or undefined. Besides the types not reflecting this behavior, resulting
in loss of type-safety, ReadOnlyFormData wasn't correctly adhering to
the FormData interface defined by the spec.
This commit fixes the .get method, so it returns the value or null, and
fixes the .getAll method, so it always returns a list, even if there are
no values for the specified key.
See https://developer.mozilla.org/en-US/docs/Web/API/FormData.