Copy endpoint response Headers instead of setting directly
Headers that are created as part of a Response object (e.g. directly
from `fetch` or via `Response.redirect`) are not mutable in some
environments, including the browser and Cloudflare Workers. If an
endpoint returns an instance of a Response currently, it will attempt
to mutate the headers, which throws an error. This only fails in
Cloudflare, so it's hard to catch before a deployment.