Fix race condition in clicknav test utility func
The Playwright docs mention that when using page.waitForNavigation(),
it's important to use Promise.all and await the navigation first, before
triggering the event (such as a link click) that will begin navigation.
Otherwise there could be a race condition where the navigation happens
too fast, before the waitForNavigation() can get set up, resulting in
waitForNavigation() timing out and the test failing.

It's very unlikely to happen in practice, but "very unlikely" isn't the
same as "impossible", so we might as well do it the way the Playwright
docs recommend, with waitForNavigation coming first.