[fix] incorrect path to generated tsconfig.json (#4179)
When attempting to update my existing sveltekit project I  added `"extends": ".svelte-kit/tsconfig.json"` like the docs say and was met with the error `'.svelte-kit/tsconfig.json' not found.` I checked the template and the correct path was `./.svelte-kit/tsconfig.json`. Without the `./`, the project fails to build and every svelte file has an error.