Implemented throwable responses from endpoints
The idea here is that endpoints can not only return a response, but also throw one.

E.g.

```js
  // return a response from an endpoint
  export async function get() {
    return { status: 200, body: {} }
  }

  // throw a response from an endpoint
  export async function get() {
    throw { status: 401 }
  }
```

This is useful for building request validation libraries. E.g. to build an auth guard
for an endpoint without this change, you would need something like:

```js
  export async function get() {
    if (!isAuthenticated()) {
      return { status: 401 }
    }
  }
```

And with this change, you can do:

```js
  export async function get() {
    // Throws a 401 if the request is not authenticated
    verifyAuthenticated()
  }
```

This is a common pattern in other web frameworks and makes it easier to structure
your code.