ensure serialized cookie values are url-encoded
TLDR

before:

```
browser
     |
(request with url-encoded cookie) (cookie=foo%2Bbar)
     |
     v
sveltekit node app
     |
(external fetch request made with url-DEcoded cookie) (cookie=foo+bar)
     |
     v
backend service
```

after:

```
browser
     |
(request with url-encoded cookie) (cookie=foo%2Bbar)
     |
     v
sveltekit node app
     |
(external fetch request made with url-ENcoded cookie) (cookie=foo%2Bbar)
     |
     v
backend service
```

My kit app was making external `fetch` requests with cookies that were different
from the the cookies that were part of the original request to the app.

My request coming from the browser had a session cookie that was
url-encoded. My sveltekit server code would then use the kit-`fetch` to
make a request to a backend api server that uses that same session token
in the cookie. But when the request to the api left the kit app, the
cookie was not url-encoded anymore. This caused problems in my api
server because it would parse the
cookies by doing a url-decode on the values. Given my session token
could contain characters that could further url-decode, it would get
garbled when it was url-decoded by the api, because it was not
url-encoded anymore.

From my superficial web readings, there seems to be no hard standard
on the encoding of cookies. But the industry trend seems to be to
url-encode cookie values.

Given that:

a) The industry trend seems to be to url-encode cookie values.
b) This cookie.js code already url-decodes cookies on the way in.
c) "Forwarded" input cookies are different on the way out than on the
way in

I assumed the attached fix is reasonable, that cookies should just
always be url-encoded.