[fix] restart on config change (#8087)
recently sveltekit changed to passing the svelte config inline to vite-plugin-svelte. With this way, vite-plugin-svelte does not watch svelte.config.js for changes to restart the vite devserver, as configFile: false indicates it should not use it.

this PR sets up a watcher in kits dev plugin that does the same. Restarting the server is absolutely required in this case.