Add extra clarification for custom error message
From time to time people do not read the entire docs and just jump in, here they see that they can add extra properties besides `message` in the error object, but doing that will give TypeScript warnings.  This change just gives a heads-up to those needing this, that later will be explained how to define this new type.