Changes ONLy to  named actions
This appears to show more changes than I made, but I only tried to change the named actions section as the default action had a typo that left the other 2 actions inside it.  So just a few characters of a change, feel free to cut and paste just them out of this change.

Sorry my first time using GIT / or GITHUB and not a Javascript or Svelte developer (yet) but been watching all the svelte vids and progress and loving Rich and teams approach and attitude, and so tried to contribute in any tiny way I can.