fix: do not crash server if specified `ADDRESS_HEADER` is missing from a request (#10285)
This fixes #10284 by moving the check for the presence of the header specified in ADDRESS_HEADER inside the getClientAddress function. The old behavior is that non-prerendered requests without the appropriate header will instantly crash the server. The new behavior is that getClientAddress() will throw an exception, like with the other checks.