Guard against null URL on `popstate`
Prevents a race condition when attempting to restore the scroll
position on a `popstate` event:

1. The page is loaded
2. The user navigates to a second route
3. The page is reloaded
4. The user immediately navigates back before hydration is complete
5. The navigation event raises an exception comparing the current and
next url

When the exception is raised, the navigation code executed after
scroll restoration additionally is not executed, resulting in
additional issues.

This is easy to test in an iOS similator:

1. Navigate to your app
2. Click to a second page
3. Focus the URL address bar
4. Move your mouse to the back button
5. Press enter to submit the URL input
6. Press the back button with your mouse almost immediately

While this specific case is fairly contrived, I believe it could also
be responsible for errors we are seeing as a user navigates back and
forward very quickly.

Related to:
https://github.com/sveltejs/kit/issues/10512