chore: check in types (#11229)
* chore: check in types

To ensure that changes to code/types doesn't result in unwanted changes in type generation, or that bumps to dts-buddy don't cause unwanted regressions, we're checking in the generated types. Types should be committed as-is (don't format it with prettier!). CI is enhanced to check that git sees no changed files after generating the types, which would mean types have changed.

* see if this fails

* nice it works

* this is unnecessary now

* update contributing

* prettier formatting getting in the way

* Update packages/kit/.prettierignore

Co-authored-by: Ben McCann <322311+benmccann@users.noreply.github.com>

* Update .github/workflows/ci.yml

Co-authored-by: Rich Harris <richard.a.harris@gmail.com>

---------

Co-authored-by: Rich Harris <richard.a.harris@gmail.com>
Co-authored-by: Ben McCann <322311+benmccann@users.noreply.github.com>