fix: skip broken files/symlinks in list_files function
The list files function walks through a directory tree and finds/filters
all files. If it encounters a symlink, it failed with an exception:

```
ENOENT: no such file or directory, stat ...
```

That can be cumbersome, because if one of your files in `static/` is a
broken symlink you are not able to test your app anymore via e.g. `npm
run dev`.