docs: update requirements for generated types
Clarifies that Typescript is required as a
dependency for generated types to be created.
When the dependency is missing they are silently
omitted. This requirement may not be obvious to
those who are running a vanilla JavaScript project
and have not used the scaffolding script (e.g.
because they migrated from an earlier version of
SvelteKit).