fix(enhanced-img): ensure ambient modules use ambient imports
Top-level imports and exports automatically upgrade a TypeScript file to the module context. This is not desired in  because we want to keep this module ambient in order for the  to apply globally across all imports of .

By importing the  type inside the ambient module context, we preserve the ambient-ness of the  file.

https://www.typescriptlang.org/docs/handbook/modules/reference.html#ambient-modules

fix