fix: don't hydrate when falling back to error page
When we encouter an error while rerunning the loaders on hydration, we need to remove the contents that were server-rendered and start fresh with the root error page. Else Svelte 5 (which is more strict with this) will fail because it tries to walk an entirely different DOM tree. Svelte 4 didn't suffer from this because it just threw away the old code, but it was still wrong to have to do it in the first place.

Fixes https://github.com/sveltejs/svelte/issues/14437