fix: decode URL before navigating when hash router is enabled (#13321)
fixes #13318

This PR decodes the URL when the hash router is enabled and the user enters the URL through the browser address bar (during app initialisation and the hashchange event). We need to do this because Safari seems to encode any extra hashes in the URL (only when accessed through the file protocol?). So, entering /#/#about in the address bar becomes /#/%23about which the client router incorrectly thinks is a route instead of just an ID.