fix: correctly handle relative anchors when hash router is enabled (#13356)
fixes #13320

This PR enables relative hash links to be used with the hash router so that links such as #example get recognised as a link to the current page. We do this by assuming links that have a hash value starting with #/ links to a route while anything else is probably linking to an element with an id on the current page.