fix: default server fetch to use prerendered paths
When using server-side fetch for internal requests, if a server route
is matched from the server `manifest.js`, it gets called without making
a real HTTP request.

However, prerendered routes are not included on this list! This is fine
when routes are prerendered with `export const prerender = true;`, but
will cause issues with a non-prerendered route also matches the same
URL as any prerendered routes.

This commit adds `prerendered_routes: Set<string>` to the `manifest.js`,
which skips calling the non-prerendered route.