fix: correctly preload links on `mousedown`/`touchstart` (#13486)
fixes #13466

This PR changes the link preload behaviour so that it only ignores a link which has already triggered a preload. Previously, it would start ignoring a link the moment it was hovered on. This caused a link that only preloads on mousedown to always be ignored since a user's mouse would hover on the link first, causing it to be ignored on the subsequent mousedown event. This PR also fixes the data-sveltekit-preload tests that did not manage to catch this regression by changing the tests to use playwright mouse hover simulation instead of programmatically dispatching the mouseover event which didn't complete before firing the other events.