fix: correctly map shared entry-point CSS files during inlining (#13431)
fixes #13423

This PR fixes the case where we fail to find a server stylesheet when Vite associates a page's CSS with a different chunk instead of the component itself. This can happen when the page is imported by another file. So, we need to crawl the imports property in the manifest for CSS too.

---------

Co-authored-by: Simon H <5968653+dummdidumm@users.noreply.github.com>