fix: Always include root layout/error nodes in manifest (#13522)
Fixes #13518.

This optimization doesn't work when an app (or a specific bundle, such as the ones generated for edge functions or when split: true in Vercel) has all prerendered routes. Given that the optimization probably isn't actually saving us much, it's probably not worth it to try to find some convoluted workaround.


---------

Co-authored-by: Simon H <5968653+dummdidumm@users.noreply.github.com>