chore: chache reroute results
Reroute is supposed to be idempotent, i.e. the same income (i.e. URL) should have always the same outcome.
With the introduction of async reroute, a crucial advantage of preloading is now lost: Reroute is called on preload, then on the actual click reroute is called again. That means that if it is async and fetches from the backend, two fetches will be made, which means the actual navigation is slowed down by a few miliseconds.
Therefore we introduce a cache which solves this problem.