feat: provide `fetch` to `reroute` (#13549)
This provides `fetch` as an argument to `reroute`, so that you can get the same benefits as the `fetch` provided to `load` functions from it.
This is important (among other things) because else you might not be able to forward cookies on the server which influence where to go to.