fix: avoid setting sequential focus navigation starting point when using hash router (#13861)
This PR fixes a regression caused by #10856 which broke hash router navigation.

We can't apply the same fix for hash routing because location.replace doesn't work since the browser interprets the entire hash as the element it should focus but we're using it as a pathname replacement. element.focus() doesn't work on Safari and Ubuntu Firefox so that's moot too.