fix: correctly set the URL when navigating during a navigation (#14004)
fixes #12809

This PR ensures that the correct navigation results are computed if the user tries to navigate during an ongoing navigation.

The root cause of the issue is that the page_changed value is evaluated incorrectly when we try navigating to a page we're already navigating to. This is because the current.url is updated to the new page before navigation is completed. Therefore, if we click the same link while navigation is pending (due to a lengthy onNavigate execution), the router will think we're already on the page we're navigating to and return the wrong navigation result. We fix this by updating current.url only after all onNavigate callbacks have completed.