chore: generate `$app/types` in a more Typescript-friendly way
This resolves the "ignore me TS, this module will be available later"-hack we did in #13864: Instead of ts-ignoring the `$app/types` import, we now generate that in advance. We didn't do that previously because there's no way to do declaration merging with types - but you can do it with interfaces, so instead we now have an interface whose properties we declaration-merge, and the types just forward those. The wrinkle is that we cannot make the properties be the types directly, else you get a "not the same types" error. Instead we're using functions because they are merged into overloads.

For this to properly work across tests (not polluting each other) I had to adjust the test infrastructure a bit, now each test project is run separately. Sadly that means it takes longer, not sure why `tsc` is so slow here.