fix: prevent loops in postbuild analysis phase (#14450)
It is currently possible for a chunk to import chunks that import
itself. I'm not exactly sure where this occurs or why as I can't
reproduce it in a simple repository, but a simple fix is to prevent
importing already visited chunks.

Helps with #14444 though doesn't fix the underlying issue