fix: avoid running `init` hook during builds when there's nothing to prerender (#14464)
fixes #14347

This PR reverts the fix from #14219 which caused the init hook to run even if there's nothing to prerender. Instead of running the init hook early, we just set the env variables when looking for prerendered remote functions. This avoids any errors for missing env variables when evaluating the user's remote function modules. We only run the init hook once we've found things we need to prerender.