feat: better remote form field interactions
Shortly after merging enhanced form validation we noticed that we can do better with respects to interacting with the form, specifically `field()/issues/input`.

This removes those properties in favor of a new `fields` property which makes interacting with the form much easier. It's using a proxy under the hood.