chore: form state rework (#14771)
We now use `$state` instead of `$state.raw` for the internal representation of `value()/set(...)`. This has several advantages:
- we can get rid of the version code which became quite a bit
- you can put a `$state` object yourself into `set(...)` and then just push to it, making array operations very easy (closes #14662)
- fine-grained without any additional footwork

The reason I made this `$state.raw` originally is so that `fields.x.value()` would fire even if something within `x` would update, but honestly that's a bad reason looking back. You _can_ get that behavior by running JSON.stringify on the result, or you just access the properties you care about and it will all take care of it automatically.

The one drawback is that prototype pollution prevention became a tiny but trickier, we now have to check for a set of keys. We should consider changing `$state` in Svelte itself to also proxify `Object.create(null)`