fix: don't redirect in forks
This prevents a redirect that a remote function could do from doing a navigation in a fork. Leverages Svelte's new forking methods from https://github.com/sveltejs/svelte/pull/17217

It's implemented by putting the context into fork, pulling it out on remote function invocation to check if we're in a fork (TODO what if you have it invoked elsewhere and you just await it in a fork context?), put that into a map, and pull it out of there when a redirect occurs to check if the remote function is only called in context of that fork, and if so instead of redirecting we tell the SvelteKit router that this new route will redirect elsewhere. In the future we could follow that redirect and run it in the fork, but this is good enough for now and ties nicely into the current SvelteKit router.

Fixes #14935