Bugfix: Fix #15004 by applying MDN recommended fix when proxying Response in page_load (#15005)
* Apply MDN recommended fix when proxying Response to avoid TypeError.

On Node v24+, the undici library used for HTTP has a Response class with
a private element. Proxying the response therefore fails when calling
clone() as the method returned by the proxy cannot access the private
state element.

* Use more comprehensive Response proxy fix that makes linter happy.

By using Object.defineProperty(), we make the function identical to
introspection. Also, the linter didn't like the `...args`.

* Add page to test undici Response TypeError.

* Add generated changeset file.

* Apply suggestion from @elliott-with-the-longest-name-on-github

---------

Co-authored-by: Elliott Johnson <hello@ell.iott.dev>